PhysicsComponent as common interface for all physics components
